% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Physics.R
\name{CascoBayPhysics}
\alias{CascoBayPhysics}
\title{Instantiate a CB_Physics R6 object}
\usage{
CascoBayPhysics(
  filename = "/mnt/ecocast/coredata/cascobay/cas3_May2018.nc",
  ...
)
}
\arguments{
\item{filename}{character either a filename or OpenDAP URL}

\item{...}{other arguments passed to \code{CB_Physics$new(filename, ...)}}
}
\value{
CB_Physics R6 Reference Object
}
\description{
Instantiate a CB_Physics R6 object
}
