% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Physics.R
\name{CB_Physics}
\alias{CB_Physics}
\title{A class for navigating  FVCOM datasets}
\description{
R6 base class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{character, path or URL}

\item{\code{NC}}{ncdf4 object}

\item{\code{M}}{sf table of mesh}

\item{\code{t0}}{POSIXct timestamp identifying the start time}

\item{\code{verbose}}{logical for helpful messaging}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{open_bounds}}{provides the indices for open boundary elements (open water)}

\item{\code{closed_bounds}}{provides the indices for closed boundary elements (shore)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CB_Physics$new()}}
\item \href{#method-get_crs}{\code{CB_Physics$get_crs()}}
\item \href{#method-get_time}{\code{CB_Physics$get_time()}}
\item \href{#method-random_points}{\code{CB_Physics$random_points()}}
\item \href{#method-char_size}{\code{CB_Physics$char_size()}}
\item \href{#method-mesh_metrics}{\code{CB_Physics$mesh_metrics()}}
\item \href{#method-append_kmetrics}{\code{CB_Physics$append_kmetrics()}}
\item \href{#method-append_bounds}{\code{CB_Physics$append_bounds()}}
\item \href{#method-plot}{\code{CB_Physics$plot()}}
\item \href{#method-clone}{\code{CB_Physics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$new(
  filename,
  origin = as.POSIXct("2018-05-01T00:00:00", tz = "UTC"),
  verbose = FALSE,
 
    crs = "+proj=tmerc +datum=NAD83 +lon_0=-70d10 lat_0=42d50 k=.9999666666666667 x_0=900000 y_0=0"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{character either a filename or OpenDAP URL}

\item{\code{origin}}{POSIXct timestamp indicating the start of the experiment}

\item{\code{verbose}}{logical for helpful messaging}

\item{\code{crs}}{CRS, likely a string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_crs"></a>}}
\if{latex}{\out{\hypertarget{method-get_crs}{}}}
\subsection{Method \code{get_crs()}}{
Retrieve the CRS
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$get_crs(form = c("proj", "wkt")[1])}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{character, one of 'proj', 'wkt'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CRS in the specified form
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_time"></a>}}
\if{latex}{\out{\hypertarget{method-get_time}{}}}
\subsection{Method \code{get_time()}}{
retrieve the time relative to some epoch/origin
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$get_time(origin = self$t0, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{origin}}{POSIXct the orogin of relative timestamps}

\item{\code{...}}{other arguments for \code{\link[fvcom]{fvcom_time}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_points"></a>}}
\if{latex}{\out{\hypertarget{method-random_points}{}}}
\subsection{Method \code{random_points()}}{
Retriev one or more random points
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$random_points(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{numeric, the number of points to generate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sf object of class POINT
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-char_size"></a>}}
\if{latex}{\out{\hypertarget{method-char_size}{}}}
\subsection{Method \code{char_size()}}{
find proxy for characteristic size based upon element size

NOTE if we update GEOS we could get size of inscribed circle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$char_size(what = c("min", "max", "square", "circle")[4])}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{character
\itemize{
\item min or square - side length of sqaure with equivalent area
\item max - side of square with double the area (or 1.4*sqrt(area))
\item circle - radius of circle with equivalent area
 }}

\item{\code{x}}{element mesh as sf}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
numeric characteristic size
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mesh_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-mesh_metrics}{}}}
\subsection{Method \code{mesh_metrics()}}{
generate kinematics metrics for a given mesh
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$mesh_metrics(
  ofile = file.path(dirname(self$filename), "mesh-elem-metrics.csv.gz")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ofile}}{optional file to save to (as non-spatial CSV)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mesh (invisibly) with variables added including
\itemize{
\item max_u, max_v, max_w maximum velocity by element over all sigma in m/s
\item area element area in m^2
\item char_dim characteristic dimension, for now the radius of the circle with equivalent area in m
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_kmetrics"></a>}}
\if{latex}{\out{\hypertarget{method-append_kmetrics}{}}}
\subsection{Method \code{append_kmetrics()}}{
append the kinematics info to the mesh table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$append_kmetrics(
  ifile = file.path(dirname(self$filename), "mesh-elem-metrics.csv.gz")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ifile}}{input CSV file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
mesh invisibly with appended variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_bounds"></a>}}
\if{latex}{\out{\hypertarget{method-append_bounds}{}}}
\subsection{Method \code{append_bounds()}}{
append the boundary info to the mesh table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$append_bounds()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
mesh invisibly with appended boundary variable where
\itemize{
 \item{open indicates the element is boundaed on at least one side by open water}
 \item{closed indicates the element is bounded on at least one side by shoreline}
 \item{internal indicated the element is not on the boundary}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot element mesh
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$plot(border = "#999999", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{border}}{character color for drawing outline}

\item{\code{...}}{other arguments for \code{\link[sf]{plot}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CB_Physics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
