% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot element mesh}
\usage{
plot_mesh(x, border = "#999999", ...)
}
\arguments{
\item{x}{the CascoBayPhysics object}

\item{border}{charcater color for drawing outline}

\item{...}{other arguments for plot.sfc}
}
\description{
Plot element mesh
}
