% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Physics.R
\name{random_point}
\alias{random_point}
\title{Generate  listing of one or more random points}
\usage{
random_point(CB, n = 1)
}
\arguments{
\item{CB}{CB_Physics object}

\item{n}{numeric the number of random points to generate}
}
\value{
sf POINT XYZ object with \code{n} features
}
\description{
Generate  listing of one or more random points
}
