% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_deployment}
\alias{read_deployment}
\title{Read Turner Johnson's 2021 deployment data}
\usage{
read_deployment(
  filename = system.file(file.path("extdata", "2021-10-12_deployment.csv.gz"), package
    = "cascophys"),
  form = c("tibble", "sf")[2],
  use = c("xy", "lonlat")[1]
)
}
\arguments{
\item{filename}{character, the file to read}

\item{form}{character, either 'tibble' or 'sf'}

\item{use}{character, determines if sf output uses lonlat or rectangular coords}
}
\value{
tibble, possible as sf POINT
}
\description{
Read Turner Johnson's 2021 deployment data
}
