% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{particle_track}
\alias{particle_track}
\title{Track a particle}
\usage{
particle_track(
  CB,
  P0 = CB$random_points(),
  tstep = 60,
  tmax = 3600 * 12,
  show_progress = TRUE,
  filename = c("particle_track.gpkg", NA)[2],
  overwrite = TRUE
)
}
\arguments{
\item{CB}{CB_Physics object}

\item{P0}{starting point, see \code{CB_Physics$random_point}}

\item{tstep}{numeric, seconds between each iteration}

\item{tmax}{numeric, number of seconds to run for}

\item{show_progress}{logical, if TRUE then show a progress bar}

\item{filename}{an optional filename to save the track to or NA to skip}

\item{overwrite}{logical, if TRUE allow overwriting of existing files}
}
\value{
sf object of type POINT
}
\description{
Track a particle
}
